requirements.txt
beautifulsoup4==4.13.4
diffusers==0.32.2
fastapi==0.115.12
numpy==2.2.6
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.2.1
pydantic==2.11.5
pydantic_settings==2.9.1
pymilvus==2.5.5
pymilvus_model==0.3.1
pytest==8.3.4
Requests==2.32.3
sentence_transformers==4.1.0
torch==2.5.1
transformers==4.48.0
uvicorn==0.34.2
vllm==0.6.6.post1

